Fix wait() call under Linux
There's no documentation of a wait() call that takes no arguments in any
POSIX standards documentations - everywhere says it should take a status
pointer. This uses waitpid instead, which waits for a specific PID
(which fixes an issue where multiple calls to StartProcess could end up
waiting for the wrong process to end).